Split tests for Aggregagtor and BucketHelper
